Remove redundant temp folder and unused imports
In many of the tests classes, we are creating two
temporary directories: one with the name "temp_<random_string>"
and inside it, we generate a new directory for each test with the name
"Test<Class_Name>_<random_string>".

I think we don't need the "temp_<random_string>" directory.
The only benefit I can think of is that it could contain multiple
temp folders from failing tests from a particular test run.
But even then, this is not a big bonus because the name
"temp_<random_string>" is not really descriptive from which test
run this directory was created.

Also, fixed the way we import "unittest_toolbox" and replace
"unittest_toolbox.Modified_TestCase.function(self)"
with "super().function()".

PS: Thanks to Jussi Kukkonen who noticed we are using two temp
folders per class in our tests.

Signed-off-by: Martin Vrachev <mvrachev@vmware.com>