New API: Disable irrelevant pylint warnings
Warnings E1120 and W0212 are irrelevant in one specific case, but
rule E1101 creates a lot of false-positives in tuf/metadata/api.

For example, there are 4 false-positives in this file related
to "meta" and "targets" fields not existing in "signed_dict" in
"from_dict" functions in Timestamp, Snapshot and Targets classes.

Signed-off-by: Martin Vrachev <mvrachev@vmware.com>