Remove typical library as an option
There are multiple limitations in the "typical" library to consider it
as an option for our new API which is going to be used from
variety of projects, some of which are big like PyPI.

The main problems I have with it are:
1. It's a relatively unknown libr1ary with only 111 stars on GitHub.
2. It doesn't allow for custom validators. It's focused on type checking
and it allows for some constraints on the function arguments/class
attributes, but nothing fancier than that. Adding the possibility for
custom validators is planned for future versions.
3. It's a one-man show with only one maintainer actively working on the
project which is a big red sign for PyPI.
4. It adds 4 additional dependencies pytzdata, pendulum, inflection,
and itself - typical.
5. It doesn't support python 3.6.

Signed-off-by: Martin Vrachev <mvrachev@vmware.com>