MetadataBundle: Modify state handling
A single state variable cannot really handle the case where we may want
load local metadata and may want to updater from remote -- but are not
required to do either.

Instead, make sure two rules apply:
* Metadata can only be loaded/updated if all metadata it depends on is
  loaded
* Metadata becomes immutable when anything that depends on it is loaded

So, loading a new timestamp is possible if root is loaded and snapshot
is not loaded.

Signed-off-by: Jussi Kukkonen <jkukkonen@vmware.com>