Prototype descriptors usage in Root and Signed
This experiment tries to help us envision how validation will look like
if we decide to use validation functions + descriptors.

What I found with this experiment is that python descriptors can't
handle dictionary keys validation as well as dictionary values.
More precisely, we can't validate that each of the "roles" in Root is
one of the 4 metadata types or that Root role keyids are unique.

The reason is that for example if we are to run this:
"root.roles = 3" this will invoke validation for root.roles, but
if we invoke: "root.roles["dwad"] = 3" then we are getting an element
of the dictionary roles and then assigning it a new value.

Signed-off-by: Martin Vrachev <mvrachev@vmware.com>