exceptions: Make more errors RepositoryErrors
Try to make errors derive from RepositoryError if they can be the
result of malicious or malfunctioning remote repository: The idea
here is that client can handle just RepositoryError instead of
individual errors (as it cannot do anything about any of them).

Also improve variable naming.

This is backwards compatible.

Signed-off-by: Jussi Kukkonen <jkukkonen@vmware.com>