exceptions: Add type annotations
exceptions.py is used from metadata.py: Annotating it is required if we
want to enable "no_untyped_call" in mypy.

The only change is UnsignedMetadataError: I would have removed argument
"signable" altogether but figured this is not worth breaking API. So
now "signable" is still not used for anything but it is now optional.

Signed-off-by: Jussi Kukkonen <jkukkonen@vmware.com>