Test new API: serialization tests for invalid arg
A while ago we decided that it's best to research each of the individuals
attributes one by one and identify what level of validation it needs
compared to how we use it:
https://github.com/theupdateframework/tuf/pull/1366#issuecomment-829288790.

This work is ongoing and there are a couple of commits already merged
for this:
- https://github.com/theupdateframework/tuf/commit/6c5d970799ffe1f79d02e24a8ea56afeae4e636b
- https://github.com/theupdateframework/tuf/commit/f20664d2fc967852d8e3aca7f951e4d1e7855e61
- https://github.com/theupdateframework/tuf/commit/41afb1e1342d5f8870180ba1ecea8894c03aec24

We want to be able to test the attributes validation against known bad
values.
The way we want to do that is with table testing we have added
using decorators for our metadata classes defined in New API:
https://github.com/theupdateframework/tuf/pull/1416.
This gives us an easy way to add new cases for each of the attributes and
not depend on external files.

Signed-off-by: Martin Vrachev <mvrachev@vmware.com>